script/requirements.txt
pandas
python-dotenv
xrpl-py